Extract coordinate modifications interface
