+++ types/src/main/kotlin/world/gregs/voidps/engine/map/Delta.kt
@@ -6,15 +6,15 @@ import world.gregs.voidps.engine.entity.Direction
  * Difference between two coordinates
  */
 @JvmInline
-value class Delta(val id: Long) {
+value class Delta(val id: Long) : Coordinate3D<Delta> {
 
     constructor(x: Int, y: Int, level: Int = 0) : this(id(x, y, level))
 
-    val x: Int
+    override val x: Int
         get() = x(id)
-    val y: Int
+    override val y: Int
         get() = y(id)
-    val level: Int
+    override val level: Int
         get() = level(id)
 
     fun isDiagonal() = isHorizontal() && isVertical()
@@ -25,26 +25,7 @@ value class Delta(val id: Long) {
 
     fun isVertical() = y != 0
 
-    fun add(x: Int, y: Int, level: Int = 0) = Delta(x = this.x + x, y = this.y + y, level = this.level + level)
-
-    fun addX(value: Int) = add(value, 0, 0)
-    fun addY(value: Int) = add(0, value, 0)
-    fun addLevel(value: Int) = add(0, 0, value)
-
-    fun minus(x: Int = 0, y: Int = 0, level: Int = 0) = add(-x, -y, -level)
-    fun delta(x: Int = 0, y: Int = 0, level: Int = 0) = Delta(this.x - x, this.y - y, this.level - level)
-
-    fun add(point: Tile) = add(point.x, point.y, point.level)
-    fun minus(point: Tile) = minus(point.x, point.y, point.level)
-    fun delta(point: Tile) = delta(point.x, point.y, point.level)
-
-    fun add(delta: Delta) = add(delta.x, delta.y, delta.level)
-    fun minus(delta: Delta) = minus(delta.x, delta.y, delta.level)
-    fun delta(delta: Delta) = delta(delta.x, delta.y, delta.level)
-
-    fun add(direction: Direction) = add(direction.delta)
-    fun minus(direction: Direction) = minus(direction.delta)
-    fun delta(direction: Direction) = delta(direction.delta)
+    override fun copy(x: Int, y: Int, level: Int) = Delta(x, y, level)
 
     fun toDirection(): Direction = when {
         x > 0 -> when {
