--- types/src/main/kotlin/world/gregs/voidps/engine/map/Id.kt
@@ -1,5 +1,0 @@
-package world.gregs.voidps.engine.map
-
-interface Id {
-    val id: Int
-}
